add collapse toggle, fix child path names, layout tweaks

+++ web/src/app/catalog/catalog.component.ts
@@ -59,7 +59,16 @@ export class CatalogComponent implements OnInit {
   }
   set folder(value: string) {
     this._folder = value;
-    this.UpdateItems();
+    this.updateItems();
+  }
+
+  get collapse(): boolean {
+    return this.configService.collapseCatalog;
+  }
+  set collapse(value: boolean) {
+    this.configService.collapseCatalog = value;
+    this.updateItems();
+    this.updatePage();
   }
 
   pagedItems: Array<Repository | string>;
@@ -114,7 +123,7 @@ export class CatalogComponent implements OnInit {
           .map(r => r.slice(0, r.length - 1).join('/'));
         this.paths = Array.from(new Set(paths));
 
-        this.UpdateItems();
+        this.updateItems();
 
         this.route.queryParams.subscribe(p => this.checkParams(p));
       }
@@ -129,24 +138,41 @@ export class CatalogComponent implements OnInit {
     this.pageCurrent = isNaN(p['page']) ? this.pageCurrent : Number(p['page']);
   }
 
-  private UpdateItems() {
+  path(item: string) {
+    return this.folder ? this.folder + '/' + item : item;
+  }
+
+  childCount(item: Repository | string): number {
+    if (this.isRepo(item)) {
+      return item.tags;
+    } else {
+      return this.repos.filter(r => r.name.startsWith(item as string)).length;
+    }
+  }
+
+  private updateItems() {
     if (this.repos) {
-      let paths: string[];
-      let repos: Repository[];
-      if (this.folder) {
-        paths = this.paths.filter(a => a !== this.folder && a.startsWith(this.folder));
-        repos = this.repos.filter(a => a.name.startsWith(this.folder));
+      if (this.collapse) {
+        let paths: string[];
+        let repos: Repository[];
+        if (this.folder) {
+          paths = this.paths.filter(a => a !== this.folder && a.startsWith(this.folder));
+          repos = this.repos.filter(a => a.name.startsWith(this.folder));
+        } else {
+          paths = this.paths;
+          repos = this.repos;
+        }
+
+        // further filter path list to those that aren't sub-paths of another item
+        const topPaths = paths.filter(a => !paths.some(b => a !== b && a.startsWith(b)));
+
+        const items = repos.filter(r => !topPaths.some(p => r.name.startsWith(p)));
+        this.items = [].concat(items).concat(topPaths.map(p => p.replace(this.folder + '/', '')));
+        this.items.sort(CatalogSort.sort);
       } else {
-        paths = this.paths;
-        repos = this.repos;
+        this._folder = null;
+        this.items = this.repos.sort(CatalogSort.sort);
       }
-
-      // further filter path list to those that aren't sub-paths of another item
-      const topPaths = paths.filter(a => !paths.some(b => a !== b && a.startsWith(b)));
-
-      const items = repos.filter(r => !topPaths.some(p => r.name.startsWith(p)));
-      this.items = [].concat(items).concat(topPaths);
-      this.items.sort(CatalogSort.sort);
     }
   }
 
@@ -170,6 +196,10 @@ export class CatalogComponent implements OnInit {
   }
 
   updatePage() {
+    if (this.pageCurrent > this.pageCount) {
+      this._pageCurrent = this.pageCount;
+    }
+
     if (this.pageSize > 0) {
       const start = (this.pageCurrent - 1) * this.pageSize;
       const end = start + this.pageSize;
@@ -206,22 +236,24 @@ export class CatalogComponent implements OnInit {
   }
 
   delete(repo: Repository) {
-    const name = repo.name;
-    if (!this.repoWorking[name]) {
-      if (confirm(`Delete all images and tags in repository ${repo.name}?`)) {
-        this.repoWorking[name] = 'Deleting';
-        this.catalogService.deleteRepo(name).subscribe((e) => {
-          if (isError(e)) {
-            if (e.resultCode === 405) {
-              this.repoError[name] = e.error;
+    if (repo.canDelete) {
+      const name = repo.name;
+      if (!this.repoWorking[name]) {
+        if (confirm(`Delete all images and tags in repository ${repo.name}?`)) {
+          this.repoWorking[name] = 'Deleting';
+          this.catalogService.deleteRepo(name).subscribe((e) => {
+            if (isError(e)) {
+              if (e.resultCode === 405) {
+                this.repoError[name] = e.error;
+              } else {
+                this.repoError[name] = e.message;
+              }
             } else {
-              this.repoError[name] = e.message;
+              this.repos = this.repos.filter(r => r.name !== repo.name);
+              delete this.repoWorking[name];
             }
-          } else {
-            this.repos = this.repos.filter(r => r.name !== repo.name);
-            delete this.repoWorking[name];
-          }
-        });
+          });
+        }
       }
     }
   }
